[SL-164] refactor: 그룹목록 조회 API에서 기존 Response DTO를 List로 감싼 Outer Class 추가
